Fix some sensor classes
