+++ homeassistant/components/nest/legacy/sensor.py
@@ -154,6 +154,11 @@ class NestBasicSensor(NestSensorDevice, SensorEntity):
     """Representation a basic Nest sensor."""
 
     @property
+    def unit_of_measurement(self):
+        """Return the unit the value is expressed in."""
+        return self._unit
+
+    @property
     def state(self):
         """Return the state of the sensor."""
         return self._state
@@ -182,6 +187,11 @@ class NestBasicSensor(NestSensorDevice, SensorEntity):
 
 class NestTempSensor(NestSensorDevice, SensorEntity):
     """Representation of a Nest Temperature sensor."""
+
+    @property
+    def unit_of_measurement(self):
+        """Return the unit the value is expressed in."""
+        return self._unit
 
     @property
     def state(self):
