+++ homeassistant/components/withings/sensor.py
@@ -32,6 +32,11 @@ class WithingsHealthSensor(BaseWithingsSensor, SensorEntity):
     """Implementation of a Withings sensor."""
 
     @property
+    def unit_of_measurement(self) -> str:
+        """Return the unit of measurement of this entity, if any."""
+        return self._attribute.unit_of_measurement
+
+    @property
     def state(self) -> None | str | int | float:
         """Return the state of the entity."""
         return self._state_data
