+++ homeassistant/components/xbee/__init__.py
@@ -9,6 +9,7 @@ import xbee_helper.const as xb_const
 from xbee_helper.device import convert_adc
 from xbee_helper.exceptions import ZigBeeException, ZigBeeTxFailure
 
+from homeassistant.components.sensor import SensorEntity
 from homeassistant.const import (
     CONF_ADDRESS,
     CONF_DEVICE,
@@ -365,7 +366,7 @@ class XBeeDigitalOut(XBeeDigitalIn):
         self._state = self._config.state2bool[pin_state]
 
 
-class XBeeAnalogIn(Entity):
+class XBeeAnalogIn(SensorEntity):
     """Representation of a GPIO pin configured as an analog input."""
 
     def __init__(self, config, device):
